backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
pydantic
langchain-ollama
pydub
TTS
ffmpeg-python
elevenlabs
aiofiles
numpy